+++ public/rename_file_event.js
@@ -1,0 +1,6 @@
+export class RenameFileEvent extends CustomEvent {
+
+    constructor(name, callback, args) {
+        super('File:rename', {detail: {name: name, callback: callback, args: args}});
+    }
+}